feat(login): save user with id from server
The `saveUserToRemote` function now returns a `Result<User>` instead of `Result<Boolean>`.
This allows the server-generated user ID to be saved locally after a successful remote save.
The `LoginViewModel` has been updated to reflect this change, saving the user with the server-provided ID to local storage.
The `User` data class now includes an `id` field.